+++ modules/cri/cri-o/00-variables_defaults.tf
@@ -1,4 +1,4 @@
-variable "crio_version" {
+variable "kubernetes_version" {
   type        = string
   description = "Version to install"
   default     = "1.29"