+++ modules/nodes/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features & Changes
 
 - Moved to opentofu. The last version tested on terraform (1.5.5) is v8.0.0.
+- Updated falco installation method
 
 ### Bugfixes